Updates tests for tdb600. Moves lockingCap tests before tbd600.
