+++ lib/constants.js
@@ -43,6 +43,7 @@ const FEDERATION_ACTIVATION_AGE = 20;
 
 const PEGIN_REJECTION_REASONS = {
   PEGIN_CAP_SURPASSED_REASON: '1',
+  LEGACY_PEGIN_MULTISIG_SENDER_REASON: '2',
   PEGIN_V1_INVALID_PAYLOAD_REASON: '4'
 };
