+++ tests/01_07_50-activate-tbd600_fork.js
@@ -1,0 +1,6 @@
+const activateForkTest = require('../lib/tests/activate-fork');
+
+activateForkTest.execute(
+    Runners.common.forks.tbd600,
+    () => Runners.hosts.federates
+);
